requirements.txt
asgiref==3.3.4
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coverage==5.5
ecdsa==0.17.0
fastapi==0.65.1
h11==0.12.0
idna==2.10
iniconfig==1.1.1
mongoengine==0.23.1
mongomock==3.22.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyasn1==0.4.8
pydantic==1.8.2
pymongo==3.11.4
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.12.1
python-jose==3.3.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.7.2
sentinels==1.0.0
six==1.16.0
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.14.0
